requirements.txt
wheel
flake8
black
django==3.1.2
tweepy==3.9.0
textblob==0.15.3
django-environ==0.4.5
